icycon/requirements.txt
Django>=3.2,<4.0
python-dotenv==1.0.0
dj-database-url==2.1.0
whitenoise==6.6.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
openai==1.3.0